Add `docs/db.json` to the root `.gitignore` so prettier will ignore it.
While this file is actually ignored by a `.gitignore` which lives inside the
`docs` directory (as is consistent with the various docs repositories or
sub-trees), prettier is being told to ignore files in the top-level
`.gitignore` and only supports a single `--ignore-path` file.

This will allow it to skip trying to format this `.json` file.