Stop running `prettier` in Travis tests.
Primarily because the results that prettier is giving on Travis are not the
same as those that are exhibited locally.  Namely, the result of running the
`npm run lint` command locally returns different files than on TravisCI.

This makes it almost entirely impossible to determine what is wrong with the
files.  It's likely enough that it's running in post-commit, but I'd be
happy to re-introduce this commit if someone can figure out what is wrong!

For now though, it's interfering with the ability to merge PRs.

As another option, perhaps CircleCI workflows would allow us to have
separate fail badges for Lint tests rather than testing the linting in every
Node.js version of the build matrix and polluting the true success and
failure of the unarguably more important tests themselves (rather than the
formatting of the code which is being tested).