Move Docs Renovate configuration into top-level `renovate.json`.
As determined in
https://github.com/renovateapp/config-help/issues/23#issuecomment-377202632,
it seems that it's not possible to have this sort of inherited
configuration.  It's possible that when I first introduced this behavior
that this behavior was on its way out (and the breaking change in Renovate
v11 on its way in).

Luckily, Renovate's shared configuration option still allows this to be
relatively painless and our `apollo-docs` shared configuration[0] does most
of the heavy lifting here so there won't be too much Renovate toe-stepping.

If this works well in the `apollo-server` repository, I'll transplant this
change to the other docs repositories[1].

[0] https://github.com/apollographql/renovate-config-apollo-docs
[1] https://github.com/meteor/hexo-theme-meteor