docs: Change links away from contentless deployment index page.
This was still linked from the v2 getting started page.  While it will
eventually link to the "Deployment" guide index page where we'll offer
guides for other deployment options, that content is not written.

For now, we'll offer those getting statred to deploy with Heroku, which is
written!

Ref: https://twitter.com/jevakallio/status/1007915638919192577
cc (and thanks) to @jevakallio for pointing this out.