(tests) Re-factor Engine tests to use `EngineMockServer` for re-usability.
While in an ideal world tests which test Engine reporting could be done
through a `nock`-like implementation (https://npm.im/nock), this slight
restructuring will make it more simple to introduce a few other Engine tests
which would otherwise conflict with this logic or necessitate duplicating
and tweaking it slightly in a number of places.

This interface should avoid the need to duplicate code with those nuances
and also DRY things up a bit.