(op-reg) Only forbid operations if `forbidUnregisteredOperations` is set.
Prior to this commit, the mode this variable describes was the default
operation for the operation registry plugin.  That said, the documentation
has already been suggesting that this be set in order to enable this
behavior.

Having the ability to toggle this behavior is certainly interesting and
useful since it helps ease into a full operation manifest mode.  While we
intend on enabling functionality in Apollo Engine to help guide this
transition, for now, the logging facilities (which are improved in recent
commits) will allow users to observe the overall behavior of their server in
terms of operation execution approval/denial.