(op-reg): Use our own in-memory store, rather than sharing the APQ cache.
Allowing the APQ cache to be used is setting up users for failure since the
APQ cache can be configured to be distributed.  This is less than ideal
since a portion of a distributed cache store might go offline rendering
operations suddenly unable to be executed, despite being previously valid.

Because of the turn over of the manifest checking, it could take a full pass
in order to re-populate these, and even that would only happen if the
manifest was actually changed (since we check the ETag).