Accommodate `apollo-server-plugin-operation-registry` with necessary tweaks.
The `apollo-server-plugin-operation-registry` was previously a part of a
monorepo which was a fork of the `apollo-server` repository which shared all
the common packages of that repository, in addition to hosting its own
`apollo-server-plugin-operation-registry` and the types which support that
package.

Its commit history has been preserved by virtue of a `git filter-branch`
command, but since that monorepo's root `package.json` hosted the
`devDependencies` (as is the setup we have with Lerna), and the root's
`types/` directory hosted some hand-crafted typings, there are some
additional `devDependencies` and `types` which have brought over from that
repository manually to support `apollo-server-plugin-operation-registry`.

Finally, the `tsconfig.json`s have been modified to avoid references to
packages which are no longer a part of the shared repository and are instead
relied on directly from npm.