Switch to `apollo-graphql` for providing the signature transformations.
As of https://github.com/apollographql/apollo-server/pull/2259, the signature
normalization transformations no longer live in `apollo-engine-reporting`.

They now live in `apollo-graphql` which allows us to drop a large portion of
the dependency tree since `apollo-engine-reporting` currently depends on
`apollo-server-core` for its type definitions.

We'll use path-based imports at the moment because not everything is
exported from the main module of `apollo-graphql` (intentionally).