Pass the `operationName` to `defaultOperationRegistrySignature`.
Start passing the `operationName` to `defaultOperationRegistrySignature`.

https://github.com/apollographql/apollo-tooling/pull/1115

If `operationName` is `null`, then the operation is anonymous.  However,
it's not possible to register anonymous operations from the `apollo` CLI.
We could fail early, however, we still want to abide by the desires of
`forbidUnregisteredOperations`, so we'll allow this hash be generated
anyway.  The hash cannot be in the manifest, so this would be okay and allow
this code to remain less conditional-y, eventually forbidding the operation
when the has is not found and `forbidUnregisteredOperations` is on.