feat(rest-datasource): support redis cluster and sentinel (#1770)
* feat(rest-datasource): add redis cluster and sentinel support

* feat(rest-datasource): update apollo-server-cache-redis to support cluster mode

* Bump minimum `engine` to Node.js 8.  Node.js 6 is now deprecated.

We're doing a major version bump here, so I'm going to take the opportunity
to "drop" Node.js 6 support.  Granted, this `engines` field is only a
suggestion that is loosely enforced by some tools (and not obeyed by Node.js
or npm), it's the spirit of the thought that counts, I think? :)

* Bump major portion of `apollo-server-cache-redis` version.

* Update changelog to reflect breaking changes instructions