Revert "chore(deps): update dependency lerna to v3.15.0 (#2810)" and pin.
In the same spirit as a2e0b68783147f45c1d3a91937a6b558cdef0e01, this reverts
commit 102ca845f70168a4a8cc036388121a11111f1b72 which Renovate re-upgraded
in #2810.

Therefore, this commit also puts a restriction within Renovate's
configuration to prevent further upgrades.  Unfortunately, this means we're
stuck at this version of Lerna until we get time to make a reproduction for
the issue I logged when this first cropped up in
https://github.com/lerna/lerna/issues/2100, but I haven't yet found the
cycles to make that happen.

I guess our repository is a reliable reproduction, right now!