Update `ApolloServer`'s `typeDefs` types to more accurately reflect reality.
While no TypeScript user would have ever been able to use it as such, Apollo
Server has for a long time (as a side-effect of its past reliance on
`graphql-tools`) accepted string types (and arrays of string types) for
`typeDefs`.

This hasn't been reflected in the TypeScript types, and has thus made it
possible to sandbag non-TypeScript users (which let's remind ourselves, is
still the majority) who were strings, when new code — like that in #2762 —
didn't defend against what TypeScript indicated would be fine.

Even if we don't necessarily support string types as a common pattern
anymore, they still work, and we need to make sure we're coding for what we
expect to work until the time comes to make a breaking change.