✓ Minimal working state! (And related reorganization)
As of this commit, the `executeOperation` method is the focal point of
Apollo Server v3 - and more importantly, after starting this new
`ApolloServer` base class to replace that which is currently provided by the
`apollo-server-core` package, IT WORKS!

This `executeOperation` method can be invoked by a transport, or tests, or
any other component.

Don't let this commit fool you: There's still a lot of work to be done, and
many bits don't work:

 * Apollo Graph Manager
 * Persisted Queries
 * Custom schema directives
 * Probably a lot more.

But for now... we celebrate, momentarily. :tada: