v3: Re-arrange `tsconfig.json` files for tests to make VSCode not complain.
There doesn't seem to have been a compilation problem, but VSCode was
insistent on not being able to find the appropriate types for
`@types/jest`-provided types like `describe`, `it`, etc.

This adjustment seems to improve that, partially in part due to providing an
explicit `jest` type to the `types` compilation configuration for tests, and
partially because VSCode seems to only recognize `tsconfig.json` files in
doing its own analysis (which is separate from that of `tsc`).