Disclose production differences in Getting Started.
While #3300 did correctly make the adjustments that are necessary when
someone's getting started experience involves a deployed version of the
application (for example, if you're deploying your Getting Started
application to Zeit's Now), the additional parameters are not otherwise
necessary.  In fact, introducing them may inadvertently cause users to
enable these, when we intentionally take the step to disable it.

This seems like an alternative approach that seems like it should hopefully
still help explain that nuance, while addressing the concern about it
becoming the de-facto copy-paste approach. :)

Cc: @trevor-scheer
Co-authored-by: Alex Homer <realalexhomer@users.noreply.github.com>
Ref: https://github.com/apollographql/apollo-server/pull/3300#issuecomment-539227138
Closes: https://github.com/apollographql/apollo-server/pull/3300