Partially merge conditional branches of overlapping concern.
NB: Suggesting side-by-side review.

As I pointed out in
https://github.com/apollographql/apollo-platform-commercial/pull/246#discussion_r346820147.

The existing logic here was branching, re-converging and then branching again
for the same `shouldForbidOperation` symbol's truthiness.  This creates a
fork in logic which seems to be for semantic reasoning, but is not.

It also made a merge conflict resolution for this branch particularly
unfriendly.  Anyhow, it's unnecessary to have this separation and this code
should read more linearly without it.