Rejigger tests to not use `willUpdateManifest`.
As allured to in 4c70889bb2dde0ba4f884297a7396f95d60adc2d, this changes the
tests to not leverage `willUpdateManifest`, which is a slightly contested
implementation that we're not certain will stay around.  Instead, this
leverages the same mechanism that `agent.test.ts` uses which was enabled by
the changes in the aforementioned commit, without mixing concerns.

And as a recap, here's more words from my other commit message which
probably belonged here anyway:

It's possible that some other internal test methods could be introduced
that allow us to control the state of the internal manifest dynamically (but
in a controlled fashion; e.g. `operationAllow('query Name { field }')`), but
until we build that, it seems reasonable to test different manifest
conditions in the same way across our different test files for this plugin
(i.e. `agent.test.ts` and `ApolloServerPluginOperationRegistry.test.ts`).