Preserve other non-`cache` properties on this APQ cache configuration.
This was already the case prior to 6d5ff773f11e9f83ed7efb15b849bd89c7375815,
however the rejiggering in that PR removed that functionality, causing any
other properties present on the cache configuration to be dropped.

The original implementation can be seen here:

https://github.com/apollographql/apollo-server/commit/6d5ff773#diff-3467ccc78fbd796513acd747ff2110a5L212-L216

This was frustrating to sort out!