Fix regression caused by #3614 when using APQ with Graph Manage… (#3638)
The fix in #3614 changed behavior which was not surfacing errors to the
extensions and request pipeline plugins when those errors occurred during
APQ negotiation.

However, it failed to consider - nor were there any tests - which ensured
that the `apollo-engine-reporting`'s mechanism didn't receive an error
earlier in the request pipeline than previously allowed.

This applies a fix which special-cases the APQ error in this case, and
avoids reporting it to Apollo Graph Manager (which is the same behavior as
before).  We also introduce a test for exercising this regression, though it's
certainly a test which makes sure that the A-E-R functionality works, not
something that tests that errors are not thrown before the expected time
in a more general "extensions"-based sense.