Add Go gqlgen lib to list of federation support (#3706)
gqlgen - one of the most popular graphql Go lib - just recently added support for federation. 
PR to the lib was merged to master Jan 7 - https://github.com/99designs/gqlgen/pull/851

The lib doesn't have documentation for the feature yet, but it's fully working, as I am already using it in one of the projects.