Only pass specific `persistedQueries` properties on `GraphQLServiceContext`.
The `persistedQueries` attribute on the GraphQLServiceContext was originally
used by the operation registry, which shared the cache with it.  This is no
longer the case.

However, while we are continuing to expand the support of the interface for
`persistedQueries`, e.g. support for default TTL values on APQ cache in
https://github.com/apollographql/apollo-server/pull/3623, we don't want to
continually expand the API surface of what we expose to the plugin API as
we're intending to deprecate its exposure in the future.

In this particular case, it certainly doesn't need to get the `ttl` default
values which are intended for APQ only.