Do not prefix potentially the same error object on every request.
Sometimes, the error object which is being caught here is in fact not a
misconfiguration of server options, but rather an error which was thrown in
a promise chain.

By appending the "Invalid options provided to ApolloServer" string to the
error object's `message` property in this method that is called on _every_
request to the server (`runHttpQuery`), we're risking appending the same
prefix to the same error over and over (i.e. "a: a: a: a: a: b").

While this prefix may have been true at one point, it is no longer possible
to enforce in a world where the schema is obtained asynchronously.