Kinda revert "Decompose `GatewayConfigBase` into types in prep..."
This reverts the decomposition introduced by commit
6b912b4b41e86bd2f7d6d67de632364eecd4fb13, but also maintains the bit that
was added right after it in 2845a8e52966f1fbba2762.

Specifically, I misjudged with the direction I was headed.  While my intent
was to have `debug` and `logger` be mutually exclusive, it turns out that in
order to preserve the existing behavior, I need to make sure that `debug:
true` is the thing that puts the default logger (`loglevel`) into the
appropriate level where it still prints out various legacy messages.

cc @trevor-scheer