Selectively import from `graphql`.
We currently have two imports for `graphql`, one which is using named
exports, and another which imports "*" - which would also get all of those
named imports.

I'm sure that the named imports are just an artifact of VSCode
auto-importing, but let's just use the named imports for their specificity
and clarity as well.