Add GraphQL Playground config section
When accessing the API endpoint generated by Serverless' deployment to AWS the returned Playground uses the /dev/graphql endpoint causing polling and server communications to fail. This can be fixed by manually changing the URL in the playground, however this can lead to confusion when the issue is not fully understood by the user.

This commit includes a new section describing how to setup GraphQL Playground to work with the serverless deployment detailed in the guide.