Place `schema` and `schemaHash` on the `GraphQLRequestContext`.
Particularly with multiple schemas in-flight at the same time, plugins
need a way to understand which schema is being operated on.

Presently, while `serverWillStart` can give you the `schema` as part of its
`GraphQLServiceContext`, there isn't (yet) an `onSchemaChange` (and I'm not
certain there will be, right now!).  Even if that did exist though, with the
introduction of the gateway, it's absolutely possible that there are
multiple overlapping requests with different schemas (even if the server
doesn't explicitly support the configuration of multiple independent
schemas right now).